remove variable `size` and parameter `has_one_per_pixel`
I wrote one special function, `MSignalCam_to_numpy`, which is
using `leaf_to_numpy()` formerly known as `v1_trick`.

This special function has a clear name, it clearly states that it
can only be used for leafes of type MSignalCam... which I like.
It uses the simple implementation of the v1_trick.

The only problem there still is, is that the v1_trick needs to be
told how many elements each event has, either 1 or 1440.
It would be great if we could ask the leaf for its "width".